@@ -1,17 +1,20 @@
 {
     "name": "laravel/laravel",
     "description": "The Laravel Framework.",
-    "keywords": ["framework", "laravel"],
+    "keywords": [
+        "framework",
+        "laravel"
+    ],
     "license": "MIT",
     "type": "project",
-    "prefer-stable" : true,
+    "prefer-stable": true,
     "minimum-stability": "dev",
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=5.6.4",
         "laravel/framework": "5.3.*",
         "jenssegers/mongodb": "^3.0",
         "predis/predis": "^1.0",
-        "laravelcollective/html": "5.3.x-dev",
+        "laravelcollective/html": "5.3.*",
         "omnipay/paypal": "*",
         "nazka/sermepa-omnipay": "dev-master",
         "creitive/laravel5-breadcrumbs": "^2.0",
@@ -35,9 +38,9 @@
     "require-dev": {
         "fzaninotto/faker": "~1.4",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~4.0",
-        "symfony/css-selector": "2.8.*|3.1.*",
-        "symfony/dom-crawler": "2.8.*|3.1.*",
+        "phpunit/phpunit": "~5.0",
+        "symfony/css-selector": "3.1.*",
+        "symfony/dom-crawler": "3.1.*",
         "barryvdh/laravel-debugbar": "^2.2",
         "barryvdh/laravel-ide-helper": "^2.2"
     },
