@@ -1,7 +1,8 @@
 <?php
 namespace App\Business\Search\Filters;
 
-interface Filter {
+interface Filter
+{
     /**
      * Apply a given search value to the builder instance.
      *
