@@ -4,7 +4,8 @@ namespace App\Business\Traits;
 
 use App\Business\Scopes\GeoIpTaxScope;
 
-trait GeoIpTrait {
+trait GeoIpTrait
+{
 
     public static function bootGeoIpTrait()
     {
