@@ -110,10 +110,10 @@ trait ImageTrait
         $mime = finfo_buffer(finfo_open(FILEINFO_MIME_TYPE), $content);
 
         // return http response
-        return new IlluminateResponse($content, 200, array(
+        return new IlluminateResponse($content, 200, [
             'Content-Type' => $mime,
             'Cache-Control' => 'Expires=' . (config('cache.image.lifetime') * 60) . ', public',
             'Etag' => md5($content)
-        ));
+        ]);
     }
 }
