@@ -9,17 +9,17 @@ trait UpdateCategoriesTrait
 {
     public static function bootUpdateCategoriesTrait()
     {
-        static::saving(function($product) {
+        static::saving(function ($product) {
             /** @var Product $product */
             $categories = [];
             /** @var Category $category */
             $category = $product->category()->first();
 
-            if(is_null($category)) {
+            if (is_null($category)) {
                 return ;
             }
 
-            if(!empty($category->father->slug)) {
+            if (!empty($category->father->slug)) {
                 $categories[] = $category->father->slug;
             }
 
