@@ -1,6 +1,6 @@
 <?php namespace App\Exceptions;
 
-
-class OutOfStockException extends \Exception {
+class OutOfStockException extends \Exception
+{
 
 }
