@@ -36,11 +36,11 @@ class UpdateCategories extends Job implements ShouldQueue
         /** @var Category $category */
         $category = $this->product->category()->first();
 
-        if(is_null($category)) {
+        if (is_null($category)) {
             return false;
         }
 
-        if(!empty($category->father->slug)) {
+        if (!empty($category->father->slug)) {
             $categories[] = $category->father->slug;
         }
 
