@@ -7,7 +7,6 @@ use App\Business\Repositories\ProductRepository;
 use App\Exceptions\OutOfStockException;
 use \Request;
 
-
 /**
  * Class Order
  * @package App
@@ -34,14 +33,14 @@ class Order extends Model
     {
         parent::boot(); // TODO: Change the autogenerated stub
 
-        self::creating(function($order) {
+        self::creating(function ($order) {
             /** @var Order $order */
             $order->cart()->map(function ($cart) {
                 /** @var Variation $variation */
                 $variation = (new ProductRepository())->findBy('_id', $cart->product_id)
                     ->productVariation(array_merge([$cart->product_id], $cart->properties));
 
-                if($variation->stock - $cart->quantity < 0 ) {
+                if ($variation->stock - $cart->quantity < 0) {
                     throw new OutOfStockException(trans('exceptions.out_of_stock', ['product' => $variation->sku]));
                 }
             });
@@ -100,7 +99,8 @@ class Order extends Model
     /**
      * @return Order|null
      */
-    public function scopeCurrentOrder($query) {
+    public function scopeCurrentOrder($query)
+    {
 
         return $query->where('token', Request::session()->get('order'));
     }
@@ -108,7 +108,7 @@ class Order extends Model
     public function conditionsFilter($condition)
     {
         return collect($this->conditions)
-            ->filter(function($conditions) use ($condition) {
+            ->filter(function ($conditions) use ($condition) {
                 return $conditions['type'] == $condition;
             })->first();
     }
