@@ -19,7 +19,7 @@ class Shipping extends Model
 
     public function states()
     {
-        return $this->country->states->first(function($state) {
+        return $this->country->states->first(function ($state) {
             return $state->_id == $this->state;
         });
     }
