@@ -1,6 +1,6 @@
 <?php
 
-return array(
+return [
 
     /*
      |--------------------------------------------------------------------------
@@ -75,4 +75,4 @@ return array(
      */
 
     'locales' => ['en', 'es'],
-);
+];
