@@ -1,6 +1,6 @@
 <?php
 
-return array(
+return [
     'dsn' => env('SENTRY_DSN'),
 
     // capture release as git sha
@@ -8,4 +8,4 @@ return array(
 
     // Capture bindings on SQL queries
     'breadcrumbs.sql_bindings' => true,
-);
+];
