@@ -12,7 +12,7 @@ class CreateCategories extends Migration
      */
     public function up()
     {
-        Schema::create('categories', function(Blueprint $table) {
+        Schema::create('categories', function (Blueprint $table) {
             $table->unique('slug');
         });
     }
