@@ -12,7 +12,7 @@ class CreateOrderTable extends Migration
      */
     public function up()
     {
-        Schema::create('orders', function(Blueprint $blueprint) {
+        Schema::create('orders', function (Blueprint $blueprint) {
         });
     }
 
