@@ -3,7 +3,8 @@
 use App\Business\Traits\Tests\ProductTrait;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 
-class CartControllerTest extends TestCase {
+class CartControllerTest extends TestCase
+{
 
     use ProductTrait, DatabaseMigrations;
 
